KALYPSO-2563 Random crash after backup
Sometimes a closed file handle is resued for a socket connection,
the binary will crash when we try to do cancleIo on that handle.

This problem is happening randomly, hard to locate where the real
root cause is.

This change is a workround to keep our backup run smooth. A separate
change will be opened to track this problem.

Change-Id: I2e28b2131bbc36bdde55ba44b39501075184601b
Signed-off-by: Felix Sun <hua.sun@emc.com>
Reviewed-on: https://gerrit.mozy.lab.emc.com/115008
Reviewed-by: Andrew Skowronski <Andrew.Skowronski@emc.com>
Reviewed-by: Joseph Cline <Joseph.Cline@dell.com>