Cmake build script - case insensitive argument
For cmake and osx Debug and Release are expected but
bash script was only working with "debug" and "release".

Reviewed over slack with kevin harris.  Another
alternative is global setting that could be
set at the top of the script, but that is a bit risker for future
break if code is moved around.

Change-Id: If6cc16cd7305802d0923c6fbaea901bf703c1b30